requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
openai==0.28.0
flask-socketio==5.1.2
python-socketio==5.5.2
python-engineio==4.3.1
flask-cors==5.0.1
simple-websocket==0.10.0